Clean GenAPI code, move tests into separate files
