+++ src/GenAPI/Microsoft.DotNet.GenAPI/AssemblySymbolOrderer.cs
@@ -18,20 +18,16 @@ namespace Microsoft.DotNet.GenAPI
         /// </summary>
         /// <param name="namespaces">List of namespaces to be sorted.</param>
         /// <returns>Returns namespaces in sorted order.</returns>
-        public static IEnumerable<INamespaceSymbol> Order(this IEnumerable<INamespaceSymbol> namespaces)
-        {
-            return namespaces.OrderBy(s => s.GetDocumentationCommentId(), StringComparer.OrdinalIgnoreCase);
-        }
+        public static IEnumerable<INamespaceSymbol> Order(this IEnumerable<INamespaceSymbol> namespaces) =>
+            namespaces.OrderBy(s => s.GetDocumentationCommentId(), StringComparer.OrdinalIgnoreCase);
 
         /// <summary>
         /// Sorts the elements of a ITypeSymbol.
         /// </summary>
         /// <param name="namespaces">List of TypeMembers to be sorted.</param>
         /// <returns>Returns TypeMembers in sorted order.</returns>
-        public static IEnumerable<T> Order<T>(this IEnumerable<T> symbols) where T : ITypeSymbol
-        {
-            return symbols.OrderBy(s => s.GetDocumentationCommentId(), StringComparer.OrdinalIgnoreCase);
-        }
+        public static IEnumerable<T> Order<T>(this IEnumerable<T> symbols) where T : ITypeSymbol =>
+            symbols.OrderBy(s => s.GetDocumentationCommentId(), StringComparer.OrdinalIgnoreCase);
 
         /// <summary>
         /// Sorts the elements of a ISymbol.
@@ -48,9 +44,8 @@ namespace Microsoft.DotNet.GenAPI
             return members.OrderBy(s => (GetMemberOrder(s), s.GetDocumentationCommentId(), StringComparer.OrdinalIgnoreCase));
         }
 
-        private static int GetMemberOrder(ISymbol symbol)
-        {
-            return symbol switch
+        private static int GetMemberOrder(ISymbol symbol) =>
+            symbol switch
             {
                 IFieldSymbol fieldSymbol when fieldSymbol.ContainingType.TypeKind == TypeKind.Enum
                     => (int)Convert.ToInt64(fieldSymbol.ConstantValue),
@@ -61,6 +56,5 @@ namespace Microsoft.DotNet.GenAPI
                 IMethodSymbol => 4,
                 _ => 5
             };
-        }
     }
 }
