+++ src/GenAPI/Microsoft.DotNet.GenAPI/AttributeDataExtensions.cs
@@ -13,39 +13,27 @@ namespace Microsoft.DotNet.GenAPI
         public static bool IsObsoleteWithUsageTreatedAsCompilationError(this AttributeData attribute)
         {
             INamedTypeSymbol? attributeClass = attribute.AttributeClass;
-            if (attributeClass == null) return false;
+            if (attributeClass is null)
+                return false;
 
             if (attributeClass.ToDisplayString() != typeof(ObsoleteAttribute).FullName)
-            {
                 return false;
-            }
 
-            if (attribute.ConstructorArguments.Count() != 2)
-            {
+            if (attribute.ConstructorArguments.Length != 2)
                 return false;
-            }
 
             TypedConstant messageArgument = attribute.ConstructorArguments.ElementAt(0);
             TypedConstant errorArgument = attribute.ConstructorArguments.ElementAt(1);
             if (messageArgument.Value == null || errorArgument.Value == null)
-            {
                 return false;
-            }
 
-            if (!(messageArgument.Value is string && errorArgument.Value is bool))
-            {
+            if (messageArgument.Value is not string || errorArgument.Value is not bool)
                 return false;
-            }
 
             return (bool)errorArgument.Value;
         }
 
-        public static bool IsDefaultMemberAttribute(this AttributeData attribute)
-        {
-            INamedTypeSymbol? attributeClass = attribute.AttributeClass;
-            if (attributeClass == null) return false;
-
-            return attributeClass.ToDisplayString() == typeof(DefaultMemberAttribute).FullName;
-        }
+        public static bool IsDefaultMemberAttribute(this AttributeData attribute) =>
+             attribute.AttributeClass?.ToDisplayString() == typeof(DefaultMemberAttribute).FullName;
     }
 }
