+++ src/GenAPI/Microsoft.DotNet.GenAPI/SyntaxGeneratorExtensions.cs
@@ -9,8 +9,6 @@ using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using System.Collections.Generic;
 using Microsoft.DotNet.ApiSymbolExtensions.Filtering;
-using System.Security.AccessControl;
-using Microsoft.VisualBasic;
 
 namespace Microsoft.DotNet.GenAPI
 {
@@ -105,7 +103,7 @@ namespace Microsoft.DotNet.GenAPI
             }
         }
 
-        // Gets the list of base class and interfaces for a given symbol <see cref="INamedTypeSymbol"/>.
+        // Gets the list of base class and interfaces for a given symbol INamedTypeSymbol.
         private static BaseListSyntax? GetBaseTypeList(this SyntaxGenerator syntaxGenerator,
             INamedTypeSymbol type,
             ISymbolFilter symbolFilter)
@@ -117,7 +115,7 @@ namespace Microsoft.DotNet.GenAPI
                 baseTypes.Add(SyntaxFactory.SimpleBaseType((TypeSyntax)syntaxGenerator.TypeExpression(type.BaseType)));
             }
 
-            // includes only interfaces that were not filtered out by the given <see cref="ISymbolFilter"/> or none of TypeParameters were filtered out.
+            // includes only interfaces that were not filtered out by the given ISymbolFilter or none of TypeParameters were filtered out.
             baseTypes.AddRange(type.Interfaces
                 .Where(i => symbolFilter.Include(i) && !i.HasInaccessibleTypeArgument(symbolFilter))
                 .Select(i => SyntaxFactory.SimpleBaseType((TypeSyntax)syntaxGenerator.TypeExpression(i))));
