+++ src/GenAPI/Microsoft.DotNet.GenAPI/SyntaxRewriter/TypeForwardAttributeCSharpSyntaxRewriter.cs
@@ -1,12 +1,9 @@
 ﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-
-using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Microsoft.CodeAnalysis.Operations;
 
 namespace Microsoft.DotNet.GenAPI.SyntaxRewriter
 {
@@ -20,10 +17,8 @@ namespace Microsoft.DotNet.GenAPI.SyntaxRewriter
         public override SyntaxNode? VisitGenericName(GenericNameSyntax node)
         {
             // skip if not of `typeof(Type<A,B,C>)` format
-            if (node.Parent == null || node.Parent.Parent == null || node.Parent.Parent is not TypeOfExpressionSyntax)
-            {
+            if (node.Parent?.Parent is not TypeOfExpressionSyntax)
                 return node;
-            }
 
             TypeArgumentListSyntax typeArgumentList = node.TypeArgumentList;
             SeparatedSyntaxList<TypeSyntax> newArguments = new();
