+++ src/Microsoft.DotNet.ApiSymbolExtensions/Filtering/CompositeSymbolFilter.cs
@@ -11,14 +11,12 @@ namespace Microsoft.DotNet.ApiSymbolExtensions.Filtering
     /// Implements the composite pattern, group the list of <see cref="ISymbol"/> and interact with them
     /// the same way as a single instance of a <see cref="ISymbol"/> object.
     /// </summary>
-    public class CompositeSymbolFilter : ISymbolFilter
+    public sealed class CompositeSymbolFilter : ISymbolFilter
     {
-        public CompositeSymbolFilter() => Filters = new();
-
         /// <summary>
         /// List on inner filters.
         /// </summary>
-        public List<ISymbolFilter> Filters { get; }
+        public List<ISymbolFilter> Filters { get; } = new();
 
         /// <summary>
         /// Determines whether the <see cref="ISymbol"/> should be included.
@@ -26,17 +24,6 @@ namespace Microsoft.DotNet.ApiSymbolExtensions.Filtering
         /// <param name="symbol"><see cref="ISymbol"/> to evaluate.</param>
         /// <returns>True to include the <paramref name="symbol"/> or false to filter it out.</returns>
         public bool Include(ISymbol symbol) => Filters.All(f => f.Include(symbol));
-
-        /// <summary>
-        /// Construct and add a new filter object to a list of filters.
-        /// </summary>
-        /// <typeparam name="T"><see cref="ISymbolFilter" /> to evaluate.</typeparam>
-        /// <returns>Returns the current instance of the class.</returns>
-        public CompositeSymbolFilter Add<T>() where T : ISymbolFilter, new()
-        {
-            Filters.Add(new T());
-            return this;
-        }
 
         /// <summary>
         /// Add a filter object to a list of filters.
